# openapi.yaml
openapi: 3.1.0
info:
  title: PEG HR LMS v1 API
  version: "1.0.0"
  description: >
    HTTP-контракт системы онбординга новых сотрудников (см. SPEC.md).
    Один сервис. Все пути под /api/v1. Реализовано в prototype/.
    Ошибка любого запроса: { error: { code, message, details? }, request_id }.
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Err:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: "snake_case машинный код" }
            message: { type: string }
            details: {}
...
    ContentType:
      type: string
      enum: [video, pdf, text]
...
security:
  - bearer: []
paths:
...
  /positions:
    get: { tags: [catalog], summary: Список должностей, responses: { "200": { description: "{ items: [{id,name,trajectory_status,trajectory_id}] }" } } }
    post:
      tags: [catalog]
      summary: Создать должность (+ пустая траектория draft + блок аттестации)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string } } } } } }
      responses: { "201": { description: position }}
  /positions/{id}:
    patch:
      tags: [catalog]
      summary: Переименовать должность
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { name: { type: string } } } } } }
      responses: { "200": { description: ok } }
    delete:
      tags: [catalog]
      summary: Удалить должность (только если нет сотрудников)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok }
        "409": { $ref: '#/components/responses/Err', description: has_employees }
  /trajectories/{positionId}:
    get:
      tags: [catalog]
      summary: Полное дерево траектории (с верными ответами) + problems[]
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: >
            { id, status, pre_onboarding[], blocks[] (regular: lessons[] с material и test;
            attestation: test), problems: string[] }
  /trajectories/{positionId}/publish:
    post:
      tags: [catalog]
      summary: draft → active (проверка полноты C-2)
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "{ ok, status: active }" }
        "422": { $ref: '#/components/responses/Err', description: "trajectory_incomplete, details = список проблем" }
  /trajectories/{positionId}/unpublish:
    post:
      tags: [catalog]
      summary: active → draft
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }
  /trajectories/{positionId}/pre-onboarding:
    post:
      tags: [catalog]
      summary: Добавить материал пре-онбординга
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content_type]
              properties:
                title: { type: string }
                content_type: { $ref: '#/components/schemas/ContentType' }
                file_url: { type: [string, "null"] }
                text_body: { type: [string, "null"] }
      responses: { "201": { description: "{ id }" } }
  /trajectories/{positionId}/pre-onboarding/order:
    put:
      tags: [catalog]
      summary: Переупорядочить материалы пре-онбординга
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [ids], properties: { ids: { type: array, items: { type: string } } } } } } }
      responses: { "200": { description: ok } }
  /pre-onboarding/{id}:
    patch:
      tags: [catalog]
      summary: Изменить материал пре-онбординга
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { title: { type: string }, content_type: { $ref: '#/components/schemas/ContentType' }, file_url: { type: string }, text_body: { type: string } } } } } }
      responses: { "200": { description: ok } }
    delete:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }
  /trajectories/{positionId}/blocks:
    post:
      tags: [catalog]
      summary: Добавить блок (regular, встаёт перед аттестацией)
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title], properties: { title: { type: string } } } } } }
      responses: { "201": { description: "{ id }" } }
  /trajectories/{positionId}/blocks/order:
    put:
      tags: [catalog]
      summary: Переупорядочить regular-блоки (аттестация всегда последняя)
      parameters: [{ name: positionId, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [ids], properties: { ids: { type: array, items: { type: string } } } } } } }
      responses: { "200": { description: ok } }
  /blocks/{id}:
    patch:
      tags: [catalog]
      summary: Переименовать блок (только regular)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { title: { type: string } } } } } }
      responses: { "200": { description: ok } }
    delete:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: ok }
        "409": { $ref: '#/components/responses/Err', description: "блок аттестации удалить нельзя" }
  /blocks/{id}/lessons:
    post:
      tags: [catalog]
      summary: Добавить урок в блок
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title], properties: { title: { type: string } } } } } }
      responses:
        "201": { description: "{ id }" }
        "409": { $ref: '#/components/responses/Err', description: "в блок аттестации уроки не добавляются" }
  /blocks/{id}/lessons/order:
    put:
      tags: [catalog]
      summary: Переупорядочить уроки блока
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [ids], properties: { ids: { type: array, items: { type: string } } } } } } }
      responses: { "200": { description: ok } }
  /blocks/{id}/test:
    put:
      tags: [catalog]
      summary: Создать/обновить тест блока аттестации
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [pass_mark_pct], properties: { pass_mark_pct: { type: integer, minimum: 1, maximum: 100 } } } } } }
      responses: { "200": { description: "{ test_id }" } }
  /lessons/{id}:
    patch:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { title: { type: string } } } } } }
      responses: { "200": { description: ok } }
    delete:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }
  /lessons/{id}/material:
    put:
      tags: [catalog]
      summary: Создать/обновить материал урока
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type]
              properties:
                content_type: { $ref: '#/components/schemas/ContentType' }
                file_url: { type: [string, "null"] }
                text_body: { type: [string, "null"] }
                min_watch_pct: { type: [integer, "null"], minimum: 0, maximum: 100 }
      responses: { "200": { description: ok } }
  /lessons/{id}/test:
    put:
      tags: [catalog]
      summary: Создать/обновить тест урока
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [pass_mark_pct], properties: { pass_mark_pct: { type: integer } } } } } }
      responses: { "200": { description: "{ test_id }" } }
  /tests/{id}/questions:
    post:
      tags: [catalog]
      summary: Добавить вопрос
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, options, correct_index]
              properties:
                text: { type: string }
                options: { type: array, minItems: 2, maxItems: 6, items: { type: string } }
                correct_index: { type: integer }
      responses: { "201": { description: "{ id }" } }
  /questions/{id}:
    patch:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
                options: { type: array, items: { type: string } }
                correct_index: { type: integer }
      responses: { "200": { description: ok } }
    delete:
      tags: [catalog]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: ok } }

  # ---------------- analytics ----------------
...